if configure has changed after pull, stops
Previously, if configure had been changed in the origin repo,
the syncing would continue using the old configure. This can
cause problems such as the following: Say you have added a new
LICENSE file and an appropriate exclusion in the rsync of bootstraph.sh.
When a user does `source configure` the new LICENSE file and
configure file get pulled but the syncing continues with the old
bootstrap which doesn't have the exclusion, thus giving you a LICENSE
file in your ~. This commit gives a basic solution of exiting if the
configure file has changed after the `git pull`.