i3lock now reads location of background
Previously, the i3lock looked for a specific image in the
home directory. Now it reads a script in the home directory
which returns the image's location. This way, when we change
background, we don't need to move the image in the home directory,
but instead just give the new location to the script. This was
inspired by the way `feh --bg` works, which is awesome.