i3: one bar config block for all displays
Maintaining multiple "bar" blocks in one config for multiple displays
has two problems:

1. Maintaining different blocks for multiple monitors accorss different
machines gets a bit tedious.

2. The whole point of multiple "bar" blocks is having different
"status_command" configuraitons. However, the "status_command" gets run
for every block, even if the display does not exist on the system. This
is a waste of resources and generally silly.

The tradeoff here is that we get the same bar config, including status
command on all monitors, everywhere. Also, with dual monitors, we can't
controll where the tray_output goes.