lock.sh: reduce "pixelation"
For the pixelation effect, the image is scaled down and up by a
percentage. If the difference is too large, there is a noticable
difference between the original resolution and the final resolution. As
a result the final pixelated screenshot doesn't fit the monitors very
well.
Here we reduce the percentage by which it is downscaled and upscaled to
avoid this issue.