clipqr: use feh instead of display
The main reason for this change is because I'm getting more familiar
with feh. Also, feh enables us to set the title of the window. This way,
from i3, we only float clipqr windows and not all feh or display
windows.

Also, looks like feh is much faster for this specific use.