add tmenu
tmenu is a small script that uses dmenu to quickly search for "templates"
and copy them to the clipboard. Now that I think of it... they are not
really templates. They are just text files. A template would imply that
it is dynamic, and while it may be in the sense that you may edit
it manually after pasting, in the context of this script it is not
dynamic at all. So I might be changing the name of the script soon.
Maybe "copymenu"?