nvim remove smartindent
I don't quite understand what smartindent does exactly. It's too
magical. I *do* know that it doesn't allow me to indent lines starting
with '#' in visual mode with >>. I suspect that smartindent is used
automatically with the c filetype anyway. I'm getting rid of it and
we'll see if I start missing something.