nvim use tmuxline and powerline fonts
I wasn't using powerline fonts in the past because there was a conflict
in the past in the unicode private use space between powerline font and
the linux-libertine font. The linux-libertine took precedence and idk
yet how to reconfigure that :S.
But at least I managed to find it was the libertine font glyph that
overrode the powerline one using gucharmap.