i3, picom: fix shadows
Now that I finally switched to i3-gaps, it was time to fix the shadows.
Big thanks to @azzamsa for the pointer. Sometimes we get caught up with
the day to day stuff that we lose sight of the important things in life,
such as keeping the rice fresh. Cheers.